+++ lib/src/types/collections.ts
@@ -74,6 +74,7 @@ export interface EntityCollection<M extends Record<string, any> = any,
      * If you are specifying your collection as code, the order is the same as the
      * one you define in `properties`. Additional columns are added at the
      * end of the list, if the order is not specified.
+     * You can use this prop to hide some properties from the table view.
      * Note that if you set this prop, other ways to hide fields, like
      * `hidden` in the property definition, will not work.
      */