+++ CHANGELOG.md
@@ -9,6 +9,7 @@ provide it.
 has been replaced with `actions`.
 - Improved preview of files that are not images, videos, or audio files.
 - Form optimizations
+- Fix for reference dialog not clearing selection
 
 ## [2.0.0-beta.1] - 2023-01-18